Fix CookieExceptions thrown for invalid cookies in HttpListenerRequest.Cookies
